+++ packages/library-base/src/commonMain/kotlin/io/realm/kotlin/serializers/RealmKSerializers.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.realm.kotlin.serializers.kotlinxserializers
+package io.realm.kotlin.serializers
 
 import io.realm.kotlin.ext.asRealmObject
 import io.realm.kotlin.ext.toRealmDictionary
