+++ packages/test-base/src/commonMain/kotlin/io/realm/kotlin/entities/SerializableSample.kt
@@ -30,13 +30,13 @@ import io.realm.kotlin.ext.backlinks
 import io.realm.kotlin.ext.realmDictionaryOf
 import io.realm.kotlin.ext.realmListOf
 import io.realm.kotlin.ext.realmSetOf
-import io.realm.kotlin.serializers.kotlinxserializers.MutableRealmIntKSerializer
-import io.realm.kotlin.serializers.kotlinxserializers.RealmAnyKSerializer
-import io.realm.kotlin.serializers.kotlinxserializers.RealmDictionaryKSerializer
-import io.realm.kotlin.serializers.kotlinxserializers.RealmInstantKSerializer
-import io.realm.kotlin.serializers.kotlinxserializers.RealmListKSerializer
-import io.realm.kotlin.serializers.kotlinxserializers.RealmSetKSerializer
-import io.realm.kotlin.serializers.kotlinxserializers.RealmUUIDKSerializer
+import io.realm.kotlin.serializers.MutableRealmIntKSerializer
+import io.realm.kotlin.serializers.RealmAnyKSerializer
+import io.realm.kotlin.serializers.RealmDictionaryKSerializer
+import io.realm.kotlin.serializers.RealmInstantKSerializer
+import io.realm.kotlin.serializers.RealmListKSerializer
+import io.realm.kotlin.serializers.RealmSetKSerializer
+import io.realm.kotlin.serializers.RealmUUIDKSerializer
 import io.realm.kotlin.types.EmbeddedRealmObject
 import io.realm.kotlin.types.MutableRealmInt
 import io.realm.kotlin.types.RealmAny
